front_end/requirements.txt
asgiref==3.5.2
bleach==5.0.1
blinker==1.5
certifi==2022.6.15
charset-normalizer==2.1.0
commonmark==0.9.1
crispy-bootstrap5==0.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-markdownify==0.9.2
docutils==0.19
et-xmlfile==1.1.0
feedgenerator==2.0.0
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.2
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
pelican==4.8.0
pelican-jinja-filters==2.1.1
psycopg2==2.9.3
Pygments==2.12.0
PyJWT==2.4.0
pymdown-extensions==9.5
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rich==12.5.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tinycss2==1.1.1
titlecase==2.3
tzdata==2022.1
Unidecode==1.3.4
urllib3==1.26.10
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0

python-dotenv~=0.20.0
